backend/requirements.txt
Django==4.2.7
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
